Split labels to the widgets crate
