+++ let-engine-audio/Cargo.toml
@@ -12,3 +12,20 @@ license = "Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+
+anyhow = { workspace = true }
+thiserror = { workspace = true }
+crossbeam = { workspace = true }
+once_cell = { workspace = true }
+parking_lot = { workspace = true }
+  [dependencies.glam]
+  workspace = true
+  features = [ "bytemuck", "mint" ]
+  [dependencies.kira]
+  version = "0.8"
+
+  [dependencies.let-engine-core]
+  version = "0.11"
+  path = "../let-engine-core"
+
+
