+++ let-engine-core/src/resources/data.rs
@@ -194,76 +194,3 @@ const SQUARE: [Vertex; 4] = [
 ];
 #[allow(dead_code)]
 const SQUARE_ID: [u32; 6] = [0, 1, 2, 1, 2, 3];
-
-/// A macro that makes it easy to create circles.
-///
-/// Returns [Data] with vertices and indices.
-///
-/// Using this with a `u32` makes a circle fan with as many corners as given.
-///
-/// Using this with a `u32` and a `f64` makes a circle fan that looks like a pie with the given percentage missing.
-///
-/// ## usage:
-/// ```rust
-/// use let_engine::prelude::*;
-///
-/// let hexagon: Data = make_circle!(6); // Makes a hexagon.
-///
-/// // Makes a pie circle fan with 20 edges with the top right part missing a quarter piece.
-/// let pie: Data = make_circle!(20, 0.75);
-/// ```
-#[macro_export]
-macro_rules! make_circle {
-    ($corners:expr) => {{ // Make a full circle fan with variable edges.
-        use let_engine::prelude::{vec2, Vertex};
-        let corners = $corners;
-        let mut vertices: Vec<Vertex> = vec![];
-        let mut indices: Vec<u32> = vec![];
-        use core::f64::consts::TAU;
-        // first point in the middle
-        vertices.push(Vertex {
-            position: vec2(0.0, 0.0),
-            tex_position: vec2(0.0, 0.0),
-        });
-        // Going through the number of steps and pushing the % of one complete TAU circle to the vertices.
-        for i in 0..corners {
-            vertices.push(vert(
-                    (TAU * ((i as f64) / corners as f64)).cos() as f32,
-                    (TAU * ((i as f64) / corners as f64)).sin() as f32,
-                ));
-        }
-        // Adding the indices adding the middle point, index and index after this one.
-        for i in 0..corners - 1 { // -1 so the last index doesn't go above the total amounts of indices.
-            indices.extend([0, i + 1, i + 2]);
-        }
-        // Completing the indices by setting the last 2 indices to the last point and the first point of the circle.
-        indices.extend([0, corners, 1]);
-        Data::Dynamic { vertices, indices }
-    }};
-    ($corners:expr, $percent:expr) => {{ // Make a pie circle fan with the amount of edges and completeness of the circle.
-        use core::f64::consts::TAU;
-        use let_engine::prelude::{vec2, Vertex};
-        let corners = $corners;
-        let percent = $percent as f64;
-        let percent: f64 = percent.clamp(0.0, 1.0);
-        let mut vertices: Vec<Vertex> = vec![];
-        let mut indices: Vec<u32> = vec![];
-
-        let count = TAU * percent;
-
-        vertices.push(vert(0.0, 0.0));
-        // Do the same as last time just with +1 iterations, because the last index doesn't go back to the first circle position.
-        for i in 0..corners + 1 {
-            vertices.push(vert(
-                    (count * ((i as f64) / corners as f64)).cos() as f32,
-                    (count * ((i as f64) / corners as f64)).sin() as f32,
-                ));
-        }
-        // This time the complete iteration is possible because the last index of the circle is not the first one as in the last.
-        for i in 0..corners {
-            indices.extend([0, i + 1, i + 2]);
-        }
-
-        Data::Dynamic { vertices, indices }
-    }};
-}
