+++ let-engine-widgets/Cargo.toml
@@ -12,3 +12,22 @@ license = "Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+once_cell = { workspace = true }
+image = { workspace = true }
+anyhow = { workspace = true }
+glyph_brush = "0.7"
+parking_lot = { workspace = true }
+
+  [dependencies.let-engine-core]
+  version = "0.11"
+  path = "../let-engine-core"
+
+  [dependencies.glam]
+  workspace = true
+  features = [ "bytemuck", "mint" ]
+
+[features]
+serde = [ "glam/serde" ]
+
+rand = [ "glam/rand" ]
+fast-math = [ "glam/fast-math" ]
